+++ lib/l10n/app_localizations_es.dart
@@ -4,9 +4,9 @@ import 'app_localizations.dart';
 
 // ignore_for_file: type=lint
 
-/// The translations for English (`en`).
-class SEn extends S {
-  SEn([String locale = 'en']) : super(locale);
+/// The translations for Spanish Castilian (`es`).
+class SEs extends S {
+  SEs([String locale = 'es']) : super(locale);
 
   @override
   String get appTitle => 'Expense Snap';