+++ lib/l10n/app_localizations_ja.dart
@@ -4,9 +4,9 @@ import 'app_localizations.dart';
 
 // ignore_for_file: type=lint
 
-/// The translations for English (`en`).
-class SEn extends S {
-  SEn([String locale = 'en']) : super(locale);
+/// The translations for Japanese (`ja`).
+class SJa extends S {
+  SJa([String locale = 'ja']) : super(locale);
 
   @override
   String get appTitle => 'Expense Snap';